+++ zone/zone.hh
@@ -10,6 +10,7 @@
 
 #include "utility.hh"
 #include "list_utility.hh"
+#include "font_ids.hh"
 
 
 /**** Definitions ****/